fixing bugs

+++ models/users.js
@@ -33,7 +33,7 @@ module.exports = function(sequelize, DataTypes) {
       }
     },
     balance: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(10, 2),
       allowNull: false,
       defaultValue: 0,
       validate: {
